[fix] 레포지토리의 id 변경

+++ src/main/java/com/wayble/server/direction/repository/EdgeRepository.java
@@ -3,5 +3,5 @@ package com.wayble.server.direction.repository;
 import com.wayble.server.direction.entity.Edge;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EdgeRepository extends JpaRepository<Edge, Integer> {
+public interface EdgeRepository extends JpaRepository<Edge, Long> {
 }
