+++ src/main/java/com/wayble/server/direction/repository/NodeRepository.java
@@ -3,5 +3,5 @@ package com.wayble.server.direction.repository;
 import com.wayble.server.direction.entity.Node;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface NodeRepository extends JpaRepository<Node, Integer> {
+public interface NodeRepository extends JpaRepository<Node, Long> {
 }
